requirements.txt
flask
google-generativeai
python-dotenv
pandas
markdown
langchain-google-genai
langchain-community